py/requirements.txt
httpx
notion-client
PyYAML
click
git+https://github.com/polym/notion2md.git